requirements.txt
pandas
numpy
scikit-learn
matplotlib
plotly
pyyaml
joblib
tqdm
# optional (comment in when needed)
# xgboost
# lightgbm
# mlflow
# hydra-core
# dvc
